+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 ### Fixes
 
 - Support the new URLSession HTTP loader for automatic network instrumentation (#8845)
+- Prevent duplicate automatic HTTP spans and breadcrumbs for URLSession task wrappers (#8846)
+- Prevent breadcrumb persistence for watchdog termination events from blocking the calling thread (#8653)
 
 ## 9.26.0